requirements.txt
requests
zhdate
python-dotenv